+++ inlang/source-code/editor/src/pages/@host/@owner/@repository/Layout.tsx
@@ -533,27 +533,18 @@ function ProjectMenu() {
 	const { project, activeProject, setActiveProject, projectList, currentBranch, activeBranch } =
 		useEditorState()
 
-	const activeProjectName = () => {
-		return (
-			activeProject()?.toString().split("/").at(-2) +
-			"/" +
-			activeProject()?.toString().split("/").at(-1)
-		)
-	}
-
-	const shortenProjectName = (projectPath: string) => {
+	const shortenProjectPath = (projectPath: string) => {
 		const projectPathArray = projectPath.split("/")
 		if (projectPathArray.length > 3) {
-			return (
-				"/" +
-				projectPathArray.at(-3) +
-				"/" +
-				projectPathArray.at(-2) +
-				"/" +
-				projectPathArray.at(-1)
-			)
+			return ".../" + projectPathArray.at(-3) + "/" + projectPathArray.at(-2) + "/" + projectPathArray.at(-1)
+		} else {
+			return "." + projectPath
 		}
-		return projectPath
+	}
+
+	const projectName = (projectPath: string) => {
+		const projectPathArray = projectPath.split("/")
+		return projectPathArray.at(-1)?.replace(".inlang", "")
 	}
 
 	return (
@@ -576,20 +567,23 @@ function ProjectMenu() {
 					<div slot="prefix">
 						<IconDescription class="-ml-1 w-5 h-5" />
 					</div>
-					{activeProject() ? activeProjectName() : "project"}
+					{activeProject() ? projectName(activeProject()!) : "project"}
 				</sl-button>
 
 				<sl-menu class="w-48 min-w-fit">
 					<For each={projectList()}>
 						{(project) => (
-							<div onClick={() => setActiveProject(project.projectPath)}>
-								<sl-menu-item
-									prop:type="checkbox"
-									prop:checked={activeProject() === project.projectPath}
-								>
-									{shortenProjectName(project.projectPath)}
-								</sl-menu-item>
-							</div>
+							<sl-menu-item
+								prop:type="checkbox"
+								prop:checked={activeProject() === project.projectPath}
+								onClick={() => setActiveProject(project.projectPath)}
+								class="flex items-center"
+							>
+								{projectName(project.projectPath)}
+								<span class="text-on-surface-variant/60 ml-2 text-xs">
+									{shortenProjectPath(project.projectPath)}
+								</span>
+							</sl-menu-item>
 						)}
 					</For>
 				</sl-menu>
@@ -720,7 +714,7 @@ function LintFilter(props: { clearFunction: any }) {
 			}}
 			class="filter border-0 focus:ring-background/100 p-0 m-0 text-sm animate-blendIn"
 		>
-			<div class={"flex items-center gap-2 ml-1 mr-0"} slot="prefix">
+			<div class="flex items-center gap-2 ml-1 mr-0" slot="prefix">
 				<p class="flex-grow-0 flex-shrink-0 text-sm font-medium text-left text-on-surface-variant">
 					Lint
 				</p>
@@ -743,7 +737,7 @@ function LintFilter(props: { clearFunction: any }) {
 				</Show>
 			</div>
 			<button slot="clear-icon" class="p-0.5">
-				<IconClose class="w-4 h-4" />
+				<IconClose width={20} height={20} />
 			</button>
 
 			<div class="flex px-3 gap-2 text-sm font-medium">
@@ -838,7 +832,7 @@ function IdsFilter(props: { clearFunction: any }) {
 							props.clearFunction
 						}}
 					>
-						<IconClose />
+						<IconClose width={20} height={20} />
 					</button>
 				</Show>
 			</div>