+++ inlang/source-code/paraglide/paraglide-sveltekit/src/runtime/index.client.ts
@@ -1,0 +1,3 @@
+export { default as ParaglideJS } from "./ParaglideJS.svelte"
+export { createI18n } from "./adapter.client.js"
+export type { ParaglideLocals } from "./locals.js"