+++ inlang/source-code/message-bundle-component/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inlang/message-bundle-component
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [8e9fc0f]
+  - @inlang/sdk@0.35.9
+
 ## 0.1.16
 
 ### Patch Changes