+++ inlang/source-code/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inlang/plugin-i18next
 
+## 4.14.9
+
+### Patch Changes
+
+- Updated dependencies [8e9fc0f]
+  - @inlang/sdk@0.35.9
+
+## 4.14.8
+
+### Patch Changes
+
+- 04e804b: add human readble id tests to plugins
+
 ## 4.14.7
 
 ### Patch Changes