+++ inlang/source-code/plugins/m-function-matcher/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inlang/plugin-m-function-matcher
 
+## 0.9.32
+
+### Patch Changes
+
+- Updated dependencies [8e9fc0f]
+  - @inlang/sdk@0.35.9
+
+## 0.9.31
+
+### Patch Changes
+
+- 04e804b: add human readble id tests to plugins
+
 ## 0.9.30
 
 ### Patch Changes